python-wrapper/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-server-tester"
version = "1.0.0"
description = "Python wrapper for mcp-server-tester-sse-http-stdio NPM package"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "stgmt", email = "your-email@example.com"}
]
keywords = [
    "mcp", "model-context-protocol", "testing", "server-testing", 
    "sse", "stdio", "http", "ai-tools", "llm-tools"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9", 
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Internet :: WWW/HTTP",
]
dependencies = [
    "pyyaml>=6.0",
    "click>=8.0.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
mcp-server-tester = "mcp_server_tester.cli:main"
mcp-tester = "mcp_server_tester.cli:main"

[project.urls]
Homepage = "https://github.com/stgmt/mcp-server-tester-python"
"Bug Reports" = "https://github.com/stgmt/mcp-server-tester-python/issues"
"Source" = "https://github.com/stgmt/mcp-server-tester-python"
"NPM Package" = "https://www.npmjs.com/package/mcp-server-tester-sse-http-stdio"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=mcp_server_tester --cov-report=html --cov-report=term-missing"